chore(api-gateway): update generated types and methods

+++ src/resources/api-gateway/user-schemas/user-schemas.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import * as HostsAPI from './hosts';
 import { HostListParams, HostListResponse, HostListResponsesV4PagePaginationArray, Hosts } from './hosts';
@@ -29,7 +30,7 @@ export class UserSchemas extends APIResource {
     params: UserSchemaCreateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<UserSchemaCreateResponse> {
-    const { zone_id, ...body } = params;
+    const { zone_id = this._client.zoneId, ...body } = params;
     return (
       this._client.post(
         `/zones/${zone_id}/api_gateway/user_schemas`,
@@ -45,10 +46,20 @@ export class UserSchemas extends APIResource {
    * @deprecated Use [Schema Validation API](https://developers.cloudflare.com/api/resources/schema_validation/) instead.
    */
   list(
-    params: UserSchemaListParams,
+    params?: UserSchemaListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<OldPublicSchemasV4PagePaginationArray, OldPublicSchema>;
+  list(
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<OldPublicSchemasV4PagePaginationArray, OldPublicSchema>;
+  list(
+    params: UserSchemaListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<OldPublicSchemasV4PagePaginationArray, OldPublicSchema> {
-    const { zone_id, ...query } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { zone_id = this._client.zoneId, ...query } = params;
     return this._client.getAPIList(
       `/zones/${zone_id}/api_gateway/user_schemas`,
       OldPublicSchemasV4PagePaginationArray,
@@ -64,10 +75,19 @@ export class UserSchemas extends APIResource {
    */
   delete(
     schemaId: string,
-    params: UserSchemaDeleteParams,
+    params?: UserSchemaDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<UserSchemaDeleteResponse>;
+  delete(schemaId: string, options?: Core.RequestOptions): Core.APIPromise<UserSchemaDeleteResponse>;
+  delete(
+    schemaId: string,
+    params: UserSchemaDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<UserSchemaDeleteResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(schemaId, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return this._client.delete(`/zones/${zone_id}/api_gateway/user_schemas/${schemaId}`, options);
   }
 
@@ -82,7 +102,7 @@ export class UserSchemas extends APIResource {
     params: UserSchemaEditParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<OldPublicSchema> {
-    const { zone_id, ...body } = params;
+    const { zone_id = this._client.zoneId, ...body } = params;
     return (
       this._client.patch(`/zones/${zone_id}/api_gateway/user_schemas/${schemaId}`, {
         body,
@@ -99,10 +119,19 @@ export class UserSchemas extends APIResource {
    */
   get(
     schemaId: string,
-    params: UserSchemaGetParams,
+    params?: UserSchemaGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<OldPublicSchema>;
+  get(schemaId: string, options?: Core.RequestOptions): Core.APIPromise<OldPublicSchema>;
+  get(
+    schemaId: string,
+    params: UserSchemaGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<OldPublicSchema> {
-    const { zone_id, ...query } = params;
+    if (isRequestOptions(params)) {
+      return this.get(schemaId, {}, params);
+    }
+    const { zone_id = this._client.zoneId, ...query } = params;
     return (
       this._client.get(`/zones/${zone_id}/api_gateway/user_schemas/${schemaId}`, {
         query,
@@ -215,7 +244,7 @@ export interface UserSchemaCreateParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param: Schema file bytes
@@ -242,7 +271,7 @@ export interface UserSchemaListParams extends V4PagePaginationArrayParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Query param: Omit the source-files of schemas and only retrieve their meta-data.
@@ -259,14 +288,14 @@ export interface UserSchemaDeleteParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface UserSchemaEditParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param: Flag whether schema is enabled for validation.
@@ -278,7 +307,7 @@ export interface UserSchemaGetParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Query param: Omit the source-files of schemas and only retrieve their meta-data.
